backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
requests>=2.32,<3.0
youtube-transcript-api>=0.6,<1.0
google-api-python-client>=2.129,<3.0
openai>=1.30,<2.0
python-dotenv>=1.0,<2.0
